efficiency create update journal entries page

+++ src/containers/pages/JournalEntries/DetailJournalEntries/index.jsx
@@ -34,11 +34,9 @@ const DetailJournalEntries = (props) => {
             confirmButtonColor: '#3085d6',
             cancelButtonColor: '#d33',
             confirmButtonText: 'Yes, delete it!'
-        }).then((result) => {
-            if (result.isConfirmed) {
-                deleteTransaction()
-            }
-        })
+        }).then((result) =>
+            result.isConfirmed && deleteTransaction()
+        )
     }
     
     const deleteTransaction = async() => {
