+++ src/containers/pages/ReceiptJournal/DetailReceiptJournal/index.jsx
@@ -47,11 +47,9 @@ const DetailReceiptJournal = (props) => {
             confirmButtonColor: '#3085d6',
             cancelButtonColor: '#d33',
             confirmButtonText: 'Yes, delete it!'
-        }).then((result) => {
-            if (result.isConfirmed) {
-                deleteTransaction()
-            }
-        })
+        }).then((result) =>
+            result.isConfirmed && deleteTransaction()
+        )
     }
 
     const getContact = async (contactId) => {
